# api/internal/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package api

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	ApiKeyAuthScopes = "ApiKeyAuth.Scopes"
	BearerAuthScopes = "BearerAuth.Scopes"
)

// Defines values for QueueEntryStatus.
const (
	CALLED    QueueEntryStatus = "CALLED"
	CANCELLED QueueEntryStatus = "CANCELLED"
	COMPLETED QueueEntryStatus = "COMPLETED"
	INSERVICE QueueEntryStatus = "IN_SERVICE"
	NOSHOW    QueueEntryStatus = "NO_SHOW"
	SKIPPED   QueueEntryStatus = "SKIPPED"
	WAITING   QueueEntryStatus = "WAITING"
)

// JoinResult defines model for JoinResult.
type JoinResult struct {
	EntryId      openapi_types.UUID `json:"entryId"`
	QrUrl        string             `json:"qrUrl"`
	TicketNumber string             `json:"ticketNumber"`
}

// PublicEntry defines model for PublicEntry.
type PublicEntry struct {
	CanCancel    bool               `json:"canCancel"`
	EntryId      openapi_types.UUID `json:"entryId"`
	EtaMinutes   int                `json:"etaMinutes"`
	Position     int                `json:"position"`
	Status       QueueEntryStatus   `json:"status"`
	TicketNumber string             `json:"ticketNumber"`
}

// QueueEntry defines model for QueueEntry.
type QueueEntry struct {
	Id            openapi_types.UUID `json:"id"`
	Position      int                `json:"position"`
	Status        QueueEntryStatus   `json:"status"`
	TicketNumber  string             `json:"ticketNumber"`
	WaitingRoomId openapi_types.UUID `json:"waitingRoomId"`
}

// QueueEntryStatus defines model for QueueEntryStatus.
type QueueEntryStatus string

// PostWaitingRoomsRoomIdSwipeJSONBody defines parameters for PostWaitingRoomsRoomIdSwipe.
type PostWaitingRoomsRoomIdSwipeJSONBody struct {
	IdCardRaw string `json:"idCardRaw"`
}

// PostWaitingRoomsRoomIdSwipeJSONRequestBody defines body for PostWaitingRoomsRoomIdSwipe for application/json ContentType.
type PostWaitingRoomsRoomIdSwipeJSONRequestBody PostWaitingRoomsRoomIdSwipeJSONBody